feat: Books are searchable, books can be added
Books are now searchable and can be added to and removed from shelves on the search results screen. A book's shelf is consistent across both screens. If a book doesn't have an image, it will have a placeholder image.
If a shelved book shows up in search results it will have a label for that shelf.